+++ backend/src/main/java/ca/metricalsky/winston/entity/fetch/YouTubeRequestEntity.java
@@ -70,9 +70,10 @@ public class YouTubeRequestEntity {
     private OffsetDateTime respondedAt;
 
     public enum RequestType {
-        CHANNELS,
         ACTIVITIES,
+        CHANNELS,
         COMMENTS,
+        PLAYLIST_ITEMS,
         REPLIES,
     }
 }